requirements.txt
openai>=1.30.0
pandas
python-dotenv
flask
requests